% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-mlstr_dictionary_functions.R
\name{apply_valueType}
\alias{apply_valueType}
\title{Apply all the valueTypes of a data dictionary to a dataset}
\usage{
apply_valueType(dataset, data_dict = data_dict_create(dataset))
}
\arguments{
\item{dataset}{A character string or tibble specifying the input dataset}

\item{data_dict}{tibble of the data dictionary (automatically generated if not provided)}
}
\value{
The dataset with the data dictionary valueType
applied to each variable
}
\description{
Uses a data dictionary in the Maelstrom Research formats (with "Variables" and "Categories"
in separate tibbles and standard columns in each) to apply their valueType to a dataset in tibble format.
If no data dictionary is provided, the function will automatically evaluate the most restrictive valueType
for each variable in the dataset and apply it.
}
\details{
Must provide an input dataset in tibble format
}
\examples{
\dontrun{
# use case 1: Apply valueType without specifying a data dictionary
apply_valueType(dataset = study_TOKYO)

# use case 2: Apply valueType using a specified data dictionary
apply_valueType(dataset = study_TOKYO , data_dict = dd_TOKYO_format_maelstrom_tagged)
}

}
