% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-mlstr_file_index.R
\name{file_index_read}
\alias{file_index_read}
\title{Read, source and open objects from an index of files}
\usage{
file_index_read(
  index = tibble(),
  file_path = "^",
  file_name = "^",
  extension = "^",
  file_type = "^",
  assign = TRUE,
  .envir = parent.frame()
)
}
\arguments{
\item{index}{The index (tibble) of a folder with file locations and metadata,
either previously generated by file_index_create() or created from folder.}

\item{file_path}{A character string specifying a file path to search by.
Can be the full string or substring (matching by regex is supported)}

\item{file_name}{A character string a file name to search by.
Can be the full string or substring (matching by regex is supported).}

\item{extension}{A character string a file extention to search by.
Can be the full string or substring (matching by regex is supported).}

\item{file_type}{A character string a file type to search by.
Can be the full string or substring (matching by regex is supported).}

\item{assign}{If TRUE, the name is automatically assigned from the name
of the object read.}

\item{.envir}{The environment to use. parent.frame() by default

Can be the full string or substring (matching by regex is supported).}
}
\value{
R objects generated in the environment or R scripts. R object names are
created automatically from their file names. Otherwise return messages indicating
what objects were created, or files opened, and if any troubles occured
}
\description{
Reads all files from an mlstr_file_index tibble as R objects to generate in the
environment or R scripts to be sourced. Any other file types will be opened in
browser (html files) or in environnement. If no index tibble is provided, the
function creates one from the working directory.
}
\examples{
\dontrun{
# use case 1: create index of files in one folder
index_DEMO <- file_index_create(folder = "TEST_demo")

# use case 5: read a file indexed matching a specific file_name query (assign = FALSE)
file_index_read(index = index_DEMO, file_name = "study_MELBOURNE_1",assign = FALSE)

# use case 6: read and download files indexed matching a specific file_name query (assign = TRUE)
file_index_read(index = index_DEMO ,file_name = "study_MELBOURNE_1",assign = TRUE)
}

}
