% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-harmo_process_harmonization.R
\name{harmo_process}
\alias{harmo_process}
\title{Generate harmonized dataset(s) and annotated data processing elements file}
\usage{
harmo_process(
  dataschema,
  data_proc_elem,
  output_harmo_folder = "harmonized_datasets",
  name_report = "annotated_data_proc_elem",
  .envir = parent.frame()
)
}
\arguments{
\item{dataschema}{A character string or list of two tibbles (Variables and Categories)
identifying the DataSchema.}

\item{data_proc_elem}{A character string or tibble R object identifying the data
processing elements file.}

\item{output_harmo_folder}{A character string specifying the folder path where
the output harmonized dataset(s) and annotated data processing elements files
will be saved.}

\item{name_report}{The name of the tibble to be generated of the annotated data
processing elements.}

\item{.envir}{The environment to use. parent.frame() by default}
}
\value{
A tibble and a csv file for each study-specific dataset that has been
harmonized saved in the specified output folder.
}
\description{
Reads DataSchema and data processing elements R objects to generate harmonized dataset(s)
and annotated data processing elements file with harmonization status and errors.
The function uses DataSchema and data processing elements specifications
to process input study-specific variables into study harmonized variables.
Documentation of each data processing action is generated in the console to support
the identification of errors and correction of the data processing elements file
and objects, as needed. An annotated data processing elements file is also produced
providing harmonization status (complete/impossible) for each DataSchema variable
in each study-specific dataset, which can be used to create a summary of harmonization
potential of DataSchema variables across study-specific datasets.
}
\details{
Must provide data processing elements, DataSchema, and study-specific datasets
in the environment.
The DataSchema is the core list of variables to be generate using study-specific
variable and must be in Maelstrom Research data dictionary format (with "Variables" and
"Categories” tibbles and standard descriptors for each). The data processing
elements file contains the harmonization potential and data processing instructions
for the harmonization project to generate harmonized datasets, in the Maelstrom
Research format.
If the harmonization process contains errors, the tibbles and csv files will not
be generated to avoid any version confusion. The user can correct their data
processing element file and re-run the process.
If the user identified errors and want to correct them later, they can specify
'undetermined' in the column 'harmo_rule' of their problematic data processing
element(s). Such data processing element will not be processed.
}
\examples{
\dontrun{
# Read the files
index_DEMO <- file_index_create(folder = "DEMO")
file_index_read(index_DEMO, file_path = "study_specific_datasets")
file_index_read(index_DEMO, file_path = "dataschema")
file_index_read(index_DEMO, file_path = "data_processing_elements")

# Example 1: all the process of harmonization with data processing elements with errors
harmo_process(
  dataschema = DEMO_dataschema,
  data_proc_elem = `DEMO_data_processing_elements - with error`,
  output_harmo_folder = "DEMO/harmonized_datasets")

# The user can use the function harmo_summary() to access the errors. In the
# environment, the user will find a report attached to harmonization.
harmo_summary(annotated_data_proc_elem)
rm(annotated_data_proc_elem)

# After correction, re-run all the process of harmonization with the corrected
# data processing elements.
harmo_process(
  dataschema = DEMO_dataschema,
  data_proc_elem = `DEMO_data_processing_elements - final`,
  name_report = "final_report",
  output_harmo_folder = "DEMO/harmonized_datasets")
}

}
