% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-harmo_process_harmonization.R
\name{harmo_summary}
\alias{harmo_summary}
\title{Generate summary of data processing elements using annotated data processing elements R object}
\usage{
harmo_summary(report)
}
\arguments{
\item{report}{A tibble of the annotated data processing elements.}
}
\description{
Reads annotated data processing elements to list processes, errors, and an
overview of each rule of harmonization. The user can take this summary to help
in the correction of their data processing elements file and re run the
process of harmonization.
}
\examples{
\dontrun{
# Read the files
file_index_read(index_DEMO, file_path = "study_specific_datasets")

# Example 1: all the process of harmonization with data processing elements with errors
file_index_read(index_DEMO, file_name = "Harmo_rule_with_error")
harmo_process(
  dataschema = Harmo_dataschema,
  data_proc_elem = Harmo_rule_with_error,
  output_harmo_folder = "DEMO/harmonized_datasets")

# Example 2: use the function harmo_summary to access the errors. In the environment,
# the user will find a report attached to harmonization process.
harmo_summary(annotated_data_proc_elem)
}

}
