% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-harmo_visualize.R
\name{harmo_visual_report}
\alias{harmo_visual_report}
\title{Generate a visual report for a harmonized dataset}
\usage{
harmo_visual_report(
  dataschema,
  data_proc_elem = NULL,
  harmonized_dataset,
  group_by = "adm_study",
  to = paste0("report/"),
  out = "plotly",
  .keep_files = FALSE
)
}
\arguments{
\item{dataschema}{A tibble identifying the DataSchema to use for harmonization}

\item{data_proc_elem}{A character string or tibble R object identifying the
data processing elements file}

\item{harmonized_dataset}{A character string or tibble identifying the input dataset.}

\item{group_by}{A character string of one column in the dataset that can be
taken as a grouping column. The visual element will be grouped and displayed
by this column. Default is “adm_study” (following the Maelstrom Reaseach naming standard).}

\item{to}{A character string identifying the folder path where the bookdown report
will be saved.}

\item{out}{parameter that specifies the graphical outputs expected in the report:
can be either 'ggplot2' or 'plotly'. gglot2 renders static plots, plotly dynamic plots.}

\item{.keep_files}{Boolean value to say whether to keep the R-markdown files.}
}
\value{
A bookdown folder containing files in the specified output folder
(current folder/docs by default). To open the file in browser, open "index.html"
}
\description{
This function generates a visual report for an harmonized dataset in in an html
bookdown document, showing descriptive statistics for each harmonized variable
to facilitate assessment of input data. Statistics and figures are generated
according to their valueTypes. The variable valueTypes is automatically detected
from the dataset and its data dictionary. The output is editable (using plotly
library) or static (using ggplot2 library).
}
\details{
Must provide a harmonized dataset and DataSchema in the Maelstrom Research formats.
The DataSchema is a tibble identifying the core data to be generated
}
\examples{
\dontrun{
# Read the files. Make sure the harmonization has been processed.
index_DEMO <- file_index_create(folder = "DEMO")
file_index_read(index = index_DEMO, file_path = "dataschema")
file_index_search(index_DEMO, file_path = "data_processing_elements")
file_index_read(index_DEMO, file_path = "harmonized_datasets")

# pool data in one tibble
harmonized_study_DEMO_table_DEMO <-
  harmonized_study_MELBOURNE_table_MELBOURNE \%>\%
  bind_rows(harmonized_study_PARIS_table_PARIS) \%>\%
  bind_rows(harmonized_study_TOKYO_table_TOKYO)

# use case 1: report of harmonized_study_DEMO_table_DEMO.
harmo_visual_report(
  harmonized_dataset = harmonized_study_DEMO_table_DEMO,
  dataschema = DEMO_dataschema,
  data_proc_elem = `DEMO_data_processing_elements - work in progress`,
  to = "DEMO/reports/harmonized_DEMO"
  # group_by = "adm_study"
)

# re-index files to include new files created
index_DEMO <- file_index_create(folder = "DEMO")

# read the book down (or open in browser the file named 'index.html')
file_index_read(index_DEMO,file_path = "DEMO/reports/harmonized_DEMO/docs/index.html")

}

}
