% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{harmonizR_help}
\alias{harmonizR_help}
\title{Call the help center for full documentation}
\usage{
harmonizR_help()
}
\description{
This function is a direct call of the documentation in the repository hosting the
package. The user accesses the description of the lastest version of the package, the
vignettes, and the list of the functions.
}
\examples{
\dontrun{
# Example 1: call the help center!

harmonizR_help()

}

}
