% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-harmo_visualize.R
\name{identify_plot_type}
\alias{identify_plot_type}
\title{Generate R script for plots based on the 'viz_type' of the variable}
\usage{
identify_plot_type(dataset = NULL, data_dict, group_by = NULL, out = "plotly")
}
\arguments{
\item{dataset}{A character string or tibble specifying the input dataset}

\item{data_dict}{tibble of the data dictionary (automatically generated if not provided)}

\item{group_by}{A character string of one column in the dataset that can be
taken as a grouping column. The visual element will be grouped and displayed
by this column.}

\item{out}{parameter that specifies the graphical outputs expected in the report:
can be either 'ggplot2' or 'plotly'. gglot2 renders static plots, plotly dynamic plots.}
}
\value{
A list of two tibbles which make up the data dictionary in Maelstrom Research
format where columns plots and summary have been added to the data dictionary
provided as an input.
}
\description{
This helper function uses the visual type attributed to a variable in a data dictionary
to generate R script to generate plots in a report. This function can be used to
manually personalize the report parameters. The plots can use an additional variable
to group each variable shown by the grouping variable.
}
\details{
The user must run identify_visual_type first to run this function.
}
\examples{
\dontrun{
# Example 1: plot R stripts for iris variables.
data_dict_create(iris, categories = "Species") \%>\%
identify_visual_type(data_dict = ., dataset = iris) \%>\%
  identify_plot_type(data_dict = ., dataset = iris) \%>\% .$Variables \%>\%
  select(name,viz_type, contains("plot"),contains("summary"))
}

}
