% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-harmo_visualize.R
\name{identify_visual_type}
\alias{identify_visual_type}
\title{Identify visual type of a variable based on valueType}
\usage{
identify_visual_type(dataset, data_dict = data_dict_create(dataset))
}
\arguments{
\item{dataset}{A character string or tibble specifying the input dataset}

\item{data_dict}{tibble of the data dictionary (automatically generated if not provided)}
}
\value{
A list of two tibbles which makes up the data dictionary in Maelstrom Research format
where a column 'viz_type' has been added to the datadictionary provided as an input.
}
\description{
This helper function analyses the content of a dataset and it data dictionary
to extract the type of visualization to generate in a report. This function can
be used to manually personalize the report parameters.
}
\examples{
\dontrun{
# Example 1: viz type of iris dataset
library(tidyverse)
identify_visual_type(dataset = iris) \%>\% .$Variables \%>\% select(name,viz_type)
}

}
