% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{mlstr_download_templates}
\alias{mlstr_download_templates}
\title{Download Maelstrom templates.}
\usage{
mlstr_download_templates(
  opal = opal.login(url = getPass::getPass("Enter the url: "), username =
    getPass::getPass("Enter the username: "), password =
    getPass::getPass("Enter the password: ")),
  from,
  to = paste0(getwd())
)
}
\arguments{
\item{opal}{Opal login attributes (asked on prompt if empty)}

\item{to}{A character string of a path where the files will be placed to in Opal
(if not specified, the working directory is the default)}
}
\description{
This helper function downloads all templates from an Opal environment and writes them in a specified directory
}
\examples{
\dontrun{
# Example 1: Download all templates amd write them in a specified directory

o <- opal.login(username = 'administrator',
               password = 'password',
               url = 'https://opal-demo.maelstrom-research.org/')
2
mlstr_download_templates( opal = o , to = "DEMO")

}

}
