% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-mlstr_opal_functions.R
\name{opal_files_push}
\alias{opal_files_push}
\title{Upload files to an Opal environment}
\usage{
opal_files_push(
  opal = opal.login(url = getPass::getPass("Enter the url: "), username =
    getPass::getPass("Enter the username: "), password =
    getPass::getPass("Enter the password: ")),
  from,
  to
)
}
\arguments{
\item{opal}{Opal login attributes (asked on prompt if empty)}

\item{from}{A character string of a path where the files will be taken from in R}

\item{to}{A character string of a path where the files will be placed to in Opal}
}
\value{
Folder(s) containing files coming from the user R environment in Opal.
The path to Opal needs to be pasted with Opal absolute path.
}
\description{
These Opal functions are wrappers of functions present in opalr. For further
information, please refer to Opal documentation. https://www.obiba.org/opalr/index.html
}
\details{
The user must be allowed to interact with their Opal
}
\examples{
\dontrun{

# use case 1: place all files in a project ("home/project/") or a user ("home/administrator/")
opal_files_push(
  opal = o,
  from = "DEMO",
  to = "home/project/DEMO")

}

}
