% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-mlstr_opal_functions.R
\name{opal_tables_pull}
\alias{opal_tables_pull}
\title{Download tables and data dictionaries from an Opal environment into the R environment}
\usage{
opal_tables_pull(
  opal = opal.login(url = getPass::getPass("Enter the url: "), username =
    getPass::getPass("Enter the username: "), password =
    getPass::getPass("Enter the password: ")),
  project,
  table = c(),
  .envir = parent.frame()
)
}
\arguments{
\item{opal}{Opal login attributes (asked on prompt if empty)}

\item{project}{A character string to name the project in Opal}

\item{table}{A character string of the name the table}

\item{.envir}{The environment to use. parent.frame() by default}
}
\value{
R objets (tibbles and list of tibbles) representing tables and their
respective data dictionary.
}
\description{
These Opal functions are wrappers of functions present in opalr. For further
information, please refer to Opal documentation. https://www.obiba.org/opalr/index.html
}
\details{
The user must be allowed to interact with their Opal
}
\examples{
\dontrun{

# use case 1: download a table and its data dictionary associated
opal_tables_pull(opal = o,project = "DEMO", table = "study_PARIS")

# use case 2: download all tables and their data dictionaries associated
opal_tables_pull(opal = o,project = "DEMO")

}

}
