% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-mlstr_opal_functions.R
\name{opal_tables_push}
\alias{opal_tables_push}
\title{Upload a tibble from the R environment into an Opal environment as tables in an Opal project}
\usage{
opal_tables_push(
  opal = opal.login(url = getPass::getPass("Enter the url: "), username =
    getPass::getPass("Enter the username: "), password =
    getPass::getPass("Enter the password: ")),
  project,
  table,
  data_dict = NULL
)
}
\arguments{
\item{opal}{Opal login attributes (asked on prompt if empty)}

\item{project}{A character string to name the project in Opal}

\item{table}{tibble of the dataset}

\item{data_dict}{tibble of the data dictionary}
}
\value{
A table in Opal.
}
\description{
These Opal functions are wrappers of functions present in opalr. For further
information, please refer to Opal documentation. https://www.obiba.org/opalr/index.html
}
\details{
The user must be allowed to interact with their Opal
}
\examples{
\dontrun{


# use case: send to Opal tables with the data dictionary (not mandatory)
opal_tables_push(
  opal = o,
  project = "DEMO",
  table = study_PARIS,
  data_dict = dd_PARIS_format_maelstrom)

}

}
