% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-harmo_visualize.R
\name{study_visual_report}
\alias{study_visual_report}
\title{Generate a visual report for a study-specific dataset}
\usage{
study_visual_report(
  dataset,
  data_dict,
  group_by = NULL,
  to = paste0("reports/"),
  out = "plotly",
  .keep_files = FALSE
)
}
\arguments{
\item{dataset}{A character string or tibble R object identifying the input dataset
(in the Maelstrom Research formats). Dataset in Maelstrom Research format has
the entity (usually participant) ID as the first column.}

\item{data_dict}{A tibble identifying the data dictionary (in the Maelstrom Research formats) associated
with the dataset. Automatically generated if not provided.}

\item{group_by}{A character string of one column in the dataset that can be
taken as a grouping column. The visual element will be grouped and displayed
by this column.}

\item{to}{A character string identifying the folder path where the bookdown report
will be saved.}

\item{out}{parameter that specifies the graphical outputs expected in the report:
can be either 'ggplot2' or 'plotly'. gglot2 renders static plots, plotly dynamic plots.}

\item{.keep_files}{Boolean value to say whether to keep the R-markdown files.}
}
\value{
A bookdown folder containing files in the specified output folder
(current folder/docs by default). To open the file in browser, open "index.html"
}
\description{
This function generates a visual report for a study-specific dataset in an html
bookdown document, showing descriptive statistics for each study-specific variable
to facilitate assessment of input data. Statistics and figures are generated
according to their valueTypes. The variable valueTypes is automatically detected
from the dataset and its data dictionary. The output is editable (using plotly
library) or static (using ggplot2 library).
}
\details{
Must provide dataset and data dictionary in the Maelstrom Research formats.
If the data dictionary is not provided, a basic one is being created during the process.
}
\examples{
\dontrun{

# Create index of files in one folder and read the files
index_DEMO <- file_index_create(folder = "DEMO")
file_index_read(index = index_DEMO, file_name = "study_TOKYO")
file_index_read(index = index_DEMO, file_name = "dd_TOKYO")

# use case 1: report of demo dataset TOKYO
study_visual_report(
  dataset = study_TOKYO,
  data_dict = dd_TOKYO_format_maelstrom_tagged,
  to = "DEMO/reports/TOKYO")

# use case 2: report of demo dataset TOKYO, grouped by gndr
study_visual_report(
  dataset = study_TOKYO,
  data_dict = dd_TOKYO_format_maelstrom_tagged,
  to = "DEMO/reports/TOKYO_gndr",group_by = "gndr",out = "ggplot2")

# re-index files to include new files created
index_DEMO <- file_index_create(folder = "DEMO")

# read the book down
file_index_read(index_DEMO,file_path = "DEMO/reports/TOKYO_gndr/docs/index.html")
}

}
